objectdictionary and Dictionary class separated out of veneer.c/.h and object.c/.h
