+++ src/builtin/veneer.h
@@ -12,6 +12,7 @@
 #include "range.h"
 #include "strng.h"
 #include "array.h"
+#include "dict.h"
 #include "closure.h"
 #include "matrix.h"
 
@@ -20,7 +21,6 @@
  * --------------------------- */
 
 #define OBJECT_CLASSNAME "Object"
-#define DICTIONARY_CLASSNAME "Dictionary"
 #define FUNCTION_CLASSNAME "Function"
 #define INVOCATION_CLASSNAME "Invocation"
 #define ERROR_CLASSNAME "Error"
@@ -28,19 +28,11 @@
 #define ERROR_TAG_PROPERTY "tag"
 #define ERROR_MESSAGE_PROPERTY "message"
 
-#define DICTIONARY_KEYS_METHOD "keys"
-#define DICTIONARY_CONTAINS_METHOD "contains"
-#define DICTIONARY_REMOVE_METHOD "remove"
-#define DICTIONARY_CLEAR_METHOD "clear"
-
 #define SETINDEX_ARGS                     "SetIndxArgs"
 #define SETINDEX_ARGS_MSG                 "Setindex method expects an index and a value as arguments."
 
 #define ENUMERATE_ARGS                    "EnmrtArgs"
 #define ENUMERATE_ARGS_MSG                "Enumerate method expects a single integer argument."
-
-#define DICT_DCTKYNTFND                   "DctKyNtFnd"
-#define DICT_DCTKYNTFND_MSG               "Key not found in dictionary."
 
 #define RESPONDSTO_ARG                    "RspndsToArg"
 #define RESPONDSTO_ARG_MSG                "Method respondsto expects a single string argument or no argrument."
@@ -50,9 +42,6 @@
 
 #define ISMEMBER_ARG                      "IsMmbrArg"
 #define ISMEMBER_ARG_MSG                  "Method ismember expects a single argument."
-
-#define DICT_DCTSTARG                     "DctStArg"
-#define DICT_DCTSTARG_MSG                 "Dictionary set methods (union, intersection, difference) expect a dictionary as the argument."
 
 #define CLASS_INVK                        "ClssInvk"
 #define CLASS_INVK_MSG                    "Cannot invoke method '%s' on a class."
